pyproject.toml
[project]
name = "omnimanager"
version = "0.1.0"
description = "OmniManager - AI-powered command and app launcher"
authors = [
    { name = "Mcode4" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "PySide6>=6.5",
    "pyxdg>=0.28",
    "llama-cpp-python",
    "sentence_transformers"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
